fix: ensure reasonable word breaks in keyboard combos
